v1.9.1
Fixing casting issue in quantizer on 64bit processors


Former-commit-id: 7d9b5aa38f572eb9816ccc4eb5ad5c0b83a7e832

+++ src/ImageProcessor/Imaging/Quantizer.cs
@@ -131,7 +131,7 @@ namespace ImageProcessor.Imaging
                 for (int col = 0; col < width; col++)
                 {
                     this.InitialQuantizePixel(new Color32(sourcePixel));
-                    sourcePixel = (IntPtr)((int)sourcePixel + this.pixelSize);
+                    sourcePixel = (IntPtr)((long)sourcePixel + this.pixelSize);
                 }
 
                 // Now I have the pixel, call the FirstPassQuantize function...
